Conflicts:
	tool/common.py

+++ tool/aligner_utils.py
@@ -25,6 +25,7 @@ import tarfile
 
 from utils import logger
 from tool.common import cd
+from tool.bam_utils import bamUtils
 
 
 class alignerUtils(object):  # pylint: disable=invalid-name
@@ -58,7 +59,7 @@ class alignerUtils(object):  # pylint: disable=invalid-name
         return True
 
     @staticmethod
-    def gem_index_genome(genome_file):
+    def gem_index_genome(genome_file, index_name=None):
         """
         Create an index of the genome FASTA file with GEM. These are saved
         alongside the assembly file.
@@ -69,7 +70,10 @@ class alignerUtils(object):  # pylint: disable=invalid-name
             Location of the assembly file in the file system
 
         """
-        command_line = 'gem-indexer -i ' + genome_file + ' -o ' + genome_file
+
+        if not index_name:
+            index_name = genome_file
+        command_line = 'gem-indexer -i ' + genome_file + ' -o ' + index_name
 
         args = shlex.split(command_line)
         process = subprocess.Popen(args)
@@ -267,7 +271,7 @@ class alignerUtils(object):  # pylint: disable=invalid-name
                         f_out.write(f_in.read())
 
             shutil.rmtree(tar_file.replace('.tar.gz', ''))
-        except IOError as error:
+        except (OSError, IOError) as error:
             logger.fatal("UNTAR: I/O error({0}): {1}".format(error.errno, error.strerror))
             return False
 
@@ -306,35 +310,25 @@ class alignerUtils(object):  # pylint: disable=invalid-name
             ' '.join(params),
         ] + reads)
 
-        cmd_sort = ' '.join([
-            'samtools sort',
-            '-O bam',
-            '-o', bam_loc,
-            reads_file_1 + '.sam'
-        ])
-
         try:
             with open(reads_file_1 + '.sam', "w") as f_out:
                 logger.info("BOWTIE2 COMMAND: " + cmd_aln)
                 process = subprocess.Popen(cmd_aln, shell=True, stdout=f_out)
                 process.wait()
-        except (IOError, OSError) as msg:
+        except (OSError, IOError) as msg:
             logger.info("I/O error({0}): {1}\n{2}".format(
                 msg.errno, msg.strerror, cmd_aln))
             return False
 
-        try:
-            logger.info("BOWTIE2 COMMAND: " + cmd_sort)
-            process = subprocess.Popen(cmd_sort, shell=True)
-            process.wait()
-        except (IOError, OSError) as msg:
-            logger.info("I/O error({0}): {1}\n{2}".format(
-                msg.errno, msg.strerror, cmd_sort))
-            return False
+        bu_handle = bamUtils()
+        return_val = bu_handle.sam_to_bam(reads_file_1 + '.sam', bam_loc)
 
-        os.remove(reads_file_1 + '.sam')
+        if return_val:
+            os.remove(reads_file_1 + '.sam')
+        else:
+            logger.warn("IO error with {} to {}".format(reads_file_1 + '.sam', bam_loc))
 
-        return True
+        return return_val
 
     def _bwa_aln_sai(self, genome_file, reads_file, params, single=True):  # pylint: disable=no-self-use
         """
@@ -369,7 +363,7 @@ class alignerUtils(object):  # pylint: disable=invalid-name
             logger.info("BWA ALN COMMAND: " + cmd_aln_sai)
             process = subprocess.Popen(cmd_aln_sai, shell=True)
             process.wait()
-        except (IOError, OSError) as msg:
+        except (OSError, IOError) as msg:
             logger.info("I/O error({0}): {1}\n{2}".format(
                 msg.errno, msg.strerror, cmd_aln_sai))
 
@@ -392,41 +386,33 @@ class alignerUtils(object):  # pylint: disable=invalid-name
             genome_file, reads_file + '.sai', reads_file
         ])
 
-        cmd_sort = ' '.join([
-            'samtools sort',
-            '-O bam',
-            '-o', bam_loc,
-            reads_file + '.sam'
-        ])
-
         self._bwa_aln_sai(genome_file, reads_file, params, True)
 
-        command_lines = [cmd_samse, cmd_sort]
-
-        # print("BWA COMMAND LINES:", command_lines)
         try:
-            for command_line in command_lines:
-                logger.info("BWA ALN COMMAND: " + command_line)
-                process = subprocess.Popen(
-                    command_line, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-                process.wait()
-                proc_out, proc_err = process.communicate()
-                logger.info("BWA ALN stdout" + proc_out)
-                logger.info("BWA ALN stderr" + proc_err)
-        except (IOError, OSError) as msg:
+            logger.info("BWA ALN COMMAND: " + cmd_samse)
+            process = subprocess.Popen(
+                cmd_samse, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+            process.wait()
+            proc_out, proc_err = process.communicate()  # pylint: disable=unused-variable
+        except (OSError, IOError) as msg:
             logger.info("I/O error({0}): {1}\n{2}".format(
-                msg.errno, msg.strerror, command_line))
+                msg.errno, msg.strerror, cmd_samse))
             proc_out, proc_err = process.communicate()
-            logger.fatal("BWA ALN stdout" + proc_out)
             logger.fatal("BWA ALN stderr" + proc_err)
             return False
 
-        os.remove(reads_file + '.sam')
-        os.remove(reads_file + '.sai')
+        bu_handle = bamUtils()
+        return_val = bu_handle.sam_to_bam(reads_file + '.sam', bam_loc)
 
-        return True
+        if return_val:
+            os.remove(reads_file + '.sam')
+            os.remove(reads_file + '.sai')
+        else:
+            logger.warn("IO error with {} to {}".format(reads_file + '.sam', bam_loc))
 
-    def bwa_aln_align_reads_paired(self, genome_file, reads_file_1, reads_file_2, bam_loc, params):
+        return return_val
+
+    def bwa_aln_align_reads_paired(self, genome_file, reads_file_1, reads_file_2, bam_loc, params):  # pylint: disable=too-many-arguments
         """
         Map the reads to the genome using BWA.
 
@@ -448,15 +434,6 @@ class alignerUtils(object):  # pylint: disable=invalid-name
             reads_file_1, reads_file_2
         ])
 
-        cmd_sort = ' '.join([
-            'samtools sort',
-            '-O bam',
-            '-o', bam_loc,
-            reads_file_1 + '.sam'
-        ])
-
-        command_lines = [cmd_samse, cmd_sort]
-
         try:
             import multiprocessing
 
@@ -474,26 +451,31 @@ class alignerUtils(object):  # pylint: disable=invalid-name
 
             f1_proc.join()
             f2_proc.join()
-        except (IOError, OSError) as msg:
+        except (OSError, IOError) as msg:
             logger.info("SAI ERROR: I/O error({0}): {1}".format(
                 msg.errno, msg.strerror))
             return False
 
         try:
-            for command_line in command_lines:
-                logger.info("BWA ALN COMMAND: " + command_line)
-                process = subprocess.Popen(command_line, shell=True)
-                process.wait()
-        except (IOError, OSError) as msg:
+            logger.info("BWA ALN COMMAND: " + cmd_samse)
+            process = subprocess.Popen(cmd_samse, shell=True)
+            process.wait()
+        except (OSError, IOError) as msg:
             logger.info("I/O error({0}): {1}\n{2}".format(
-                msg.errno, msg.strerror, command_line))
+                msg.errno, msg.strerror, cmd_samse))
             return False
 
-        os.remove(reads_file_1 + '.sam')
-        os.remove(reads_file_1 + '.sai')
-        os.remove(reads_file_2 + '.sai')
+        bu_handle = bamUtils()
+        return_val = bu_handle.sam_to_bam(reads_file_1 + '.sam', bam_loc)
 
-        return True
+        if return_val:
+            os.remove(reads_file_1 + '.sam')
+            os.remove(reads_file_1 + '.sai')
+            os.remove(reads_file_2 + '.sai')
+        else:
+            logger.warn("IO error with {} to {}".format(reads_file_1 + '.sam', bam_loc))
+
+        return return_val
 
     @staticmethod
     def bwa_mem_align_reads(
@@ -521,32 +503,22 @@ class alignerUtils(object):  # pylint: disable=invalid-name
             genome_file
         ] + reads)
 
-        cmd_sort = ' '.join([
-            'samtools sort',
-            '-O bam',
-            '-o', bam_loc,
-            reads_file_1 + '.sam'
-        ])
-
         try:
             with open(reads_file_1 + '.sam', "w") as f_out:
                 logger.info("BWA MEM COMMAND: " + cmd_aln)
                 process = subprocess.Popen(cmd_aln, shell=True, stdout=f_out)
                 process.wait()
-        except (IOError, OSError) as msg:
+        except (OSError, IOError) as msg:
             logger.info("I/O error({0}): {1}\n{2}".format(
                 msg.errno, msg.strerror, cmd_aln))
             return False
 
-        try:
-            logger.info("BWA MEM COMMAND: " + cmd_sort)
-            process = subprocess.Popen(cmd_sort, shell=True)
-            process.wait()
-        except (IOError, OSError) as msg:
-            logger.info("I/O error({0}): {1}\n{2}".format(
-                msg.errno, msg.strerror, cmd_sort))
-            return False
+        bu_handle = bamUtils()
+        return_val = bu_handle.sam_to_bam(reads_file_1 + '.sam', bam_loc)
 
-        os.remove(reads_file_1 + '.sam')
+        if return_val:
+            os.remove(reads_file_1 + '.sam')
+        else:
+            logger.warn("IO error with {} to {}".format(reads_file_1 + '.sam', bam_loc))
 
-        return True
+        return return_val
